requirements.txt
streamlit==1.38.0
pillow==10.4.0pip